requirements.txt
requests
schedule
msal
python-dotenv
aiohttp